requirements.txt
certifi
charset-normalizer
einops
filelock
fsspec
hf-xet
huggingface-hub
idna
Jinja2
MarkupSafe
mpmath
networkx
numpy
opencv-python
packaging
pillow
PyYAML
regex
requests
safetensors
setuptools
sympy
timm
tokenizers
torch
torchvision
tqdm
transformers
typing_extensions
urllib3
